Add molecule exceptions to validation exceptions
If the composable function under test throws an exception, this
doesn't make it to the validation lambda until they await an event.
If the validation lambda is asserting against some side effect of
the composable function, then the test will fail and it won't be
immediately clear that the composable function threw an exception
and was torn down. This change adds any unconsumed errors to the
suppressed list on thrown validation exceptions for visibility.