Fix for hanging MoleculeConcurrentTest
This should hopefully resolve the issue for the hanging test. I believe this should achieve the desired results of verifying the Immediate clock starts on the current thread without the use of backward writes.

The existing test case had the risk of recomposing on every frame as it was writing state that had already been read within the composable. If UnconfinedTestDispatcher is undesirable, I can look into alternatives for making sure the recompositon runs predictably.